refactor(finance): reorganiza componentes de transferencia
