+++ stories/BankAccountModal.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/nextjs';
 import { within, expect, fn } from 'storybook/test';
-import BankAccountModal from '@/app/components/modals/BankAccountModal';
+import BankAccountModal from '@/app/admin/financeiro/transferencia/components/modals/BankAccountModal';
 import { AuthProvider } from '../lib/context/AuthContext';
 import { ThemeProvider } from '../lib/context/ThemeContext';
